refactor(gateway): flatten provider folder and consolidate gateway implementations
