+++ package.json
@@ -18,16 +18,6 @@
         "default": "./dist/index.cjs"
       }
     },
-    "./provider": {
-      "import": {
-        "types": "./dist/provider/index.d.ts",
-        "default": "./dist/provider/index.js"
-      },
-      "require": {
-        "types": "./dist/provider/index.d.cts",
-        "default": "./dist/provider/index.cjs"
-      }
-    },
     "./nestjs": {
       "import": {
         "types": "./dist/nestjs/index.d.ts",
