+++ src/__tests__/package-exports.spec.ts
@@ -14,9 +14,10 @@ const packageJson = JSON.parse(readFileSync(fileURLToPath(new URL('../../package
 
 describe('package exports', () => {
   it('keeps the release-critical entrypoints in the export map', () => {
-    for (const subpath of ['.', './provider', './nestjs']) {
+    for (const subpath of ['.', './nestjs']) {
       expect(packageJson.exports[subpath]).toBeDefined();
     }
+    expect(packageJson.exports['./provider']).toBeUndefined();
   });
 
   it('does not expose removed subpaths', () => {
